Getting rid of duplicate system
