--- preload-agents.js
@@ -1,41 +1,0 @@
-const { contextBridge, ipcRenderer } = require('electron');
-
-contextBridge.exposeInMainWorld('api', {
-    receiveAnalysis: (callback) => {
-        ipcRenderer.on('analysis-result', (event, data) => callback(data));
-    },
-    platform: process.platform,  // Add this line
-    
-    onAnalysisStart: (callback) => {
-        ipcRenderer.on('analysis-start', () => callback());
-    },
-
-    onAgentProgress: (callback) => {
-        ipcRenderer.on('agent-progress', (event, update) => callback(update));
-    },
-
-    onInitialSetup: (callback) => {
-        ipcRenderer.on('show-initial-setup', () => callback());
-    },
-    
-    onSetupComplete: (callback) => {
-        ipcRenderer.on('initial-setup-complete', () => callback());
-    },
-    
-    onSetupError: (callback) => {
-        ipcRenderer.on('initial-setup-error', (event, error) => callback(error));
-    },
-
-    // Server events
-    onServerStarting: (callback) => {
-        ipcRenderer.on('server-starting', () => callback());
-    },
-
-    onServerReady: (callback) => {
-        ipcRenderer.on('server-ready', () => callback());
-    },
-
-    onServerError: (callback) => {
-        ipcRenderer.on('server-error', (event, error) => callback(error));
-    },
-});
