--- src/index-agents.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Criticaide</title>
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="agent-style.css">
-</head>
-<body>
-    <main class="app-container">
-        <header class="app-header">
-            <h1><span class="criticaide-blue">Criticaide</span></h1>
-            <p class="header-subtitle" id="shortcut-text">Press Ctrl+Alt+Shift+T to analyze copied text</p>
-        </header>
-        
-        <div id="results" class="results-container" style="display: none;">
-            <!-- Content will be injected here by renderer-agents.js -->
-        </div>
-        <div id="loading" class="loading-container" style="display: flex;">
-            <!-- Loading screen will be mounted here -->
-        </div>
-
-        <footer class="app-footer">
-            <p class="shortcut-hint">Your first line of defense in an era of digital manipulation.</p>
-        </footer>
-    </main>
-    <script src="renderer-agents.js"></script>
-</body>
-</html>
