+++ src/index.html
@@ -5,12 +5,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Criticaide</title>
     <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="agent-style.css">
 </head>
 <body>
     <main class="app-container">
         <header class="app-header">
-            <h1>Criticaide</h1>
-            <p class="header-subtitle">Press Ctrl+Alt+Shift+T to analyze copied text</p>
+            <h1><span class="criticaide-blue">Criticaide</span></h1>
+            <p class="header-subtitle" id="shortcut-text">Press Ctrl+Alt+Shift+T to analyze copied text</p>
         </header>
         
         <div id="results" class="results-container" style="display: none;">
@@ -21,7 +22,7 @@
         </div>
 
         <footer class="app-footer">
-            <p class="shortcut-hint">💡 Your first line of defense in an era of digital manipulation.</p>
+            <p class="shortcut-hint">Your first line of defense in an era of digital manipulation.</p>
         </footer>
     </main>
     <script src="renderer.js"></script>
